+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkManager.cs
@@ -1468,6 +1468,20 @@ namespace Unity.Netcode
         }
 
         /// <summary>
+        /// Get the TransportId from the associated ClientId.
+        /// </summary>
+        /// <param name="clientId">The ClientId to get the TransportId from</param>
+        /// <returns>The TransportId associated with the given ClientId</returns>
+        public ulong GetTransportIdFromClientId(ulong clientId) => ConnectionManager.ClientIdToTransportId(clientId);
+
+        /// <summary>
+        /// Get the ClientId from the associated TransportId.
+        /// </summary>
+        /// <param name="transportId">The TransportId to get the ClientId from</param>
+        /// <returns>The ClientId from the associated TransportId</returns>
+        public ulong GetClientIdFromTransportId(ulong transportId) => ConnectionManager.TransportIdToClientId(transportId);
+
+        /// <summary>
         /// Disconnects the remote client.
         /// </summary>
         /// <param name="clientId">The ClientId to disconnect</param>
@@ -1593,8 +1607,9 @@ namespace Unity.Netcode
             // Completely reset the NetworkClient
             ConnectionManager.LocalClient = new NetworkClient();
 
-            // This cleans up the internal prefabs list
+            // Clean up the internal prefabs data
             NetworkConfig?.Prefabs?.Shutdown();
+            PrefabHandler.Shutdown();
 
             // Reset the configuration hash for next session in the event
             // that the prefab list changes