+++ com.unity.netcode.gameobjects/CHANGELOG.md
@@ -21,6 +21,13 @@ Additional documentation and release notes are available at [Multiplayer Documen
 ### Changed
 
 
+## [2.4.3] - 2025-06-25
+
+### Fixed
+
+- Fixed issue where spawned objects with `NetworkObject.DontDestroyWithOwner` set to `false` would not be destroyed when using a client-server network topology. (#3522)
+
+
 ## [2.4.2] - 2025-06-13
 
 ### Fixed